fix: make HooksGenerator a proper Rails generator (fixes #99)
- Move HooksGenerator to lib/generators/migration_guard/hooks/hooks_generator.rb
- Make it inherit from Rails::Generators::Base
- Use proper Rails generator conventions (class_option, say_status, Thor::Error)
- Fix exit handling to use Thor::Error instead of direct exit

The generator now works as documented:
  rails generate migration_guard:hooks
  rails generate migration_guard:hooks --pre-push

Fixes #99
